build: modularize codebase and improve type generation
Split monolithic skills-core.ts into focused modules (skills, agents,
commands, frontmatter, bootstrap, walk-dir). Remove skill-tool hook
system. Enable TypeScript declaration generation in build process.
Update all imports and tests accordingly.
