+++ src/lib/skill-tool.ts
@@ -3,105 +3,10 @@ import path from 'node:path'
 import type { ToolDefinition } from '@opencode-ai/plugin'
 import { tool } from '@opencode-ai/plugin/tool'
 import { convertContent } from './converter.js'
-import type { SkillInfo } from './skills-core.js'
-import { findSkillsInDir, stripFrontmatter } from './skills-core.js'
+import { stripFrontmatter } from './frontmatter.js'
+import type { SkillInfo } from './skills.js'
+import { findSkillsInDir, formatSkillsXml } from './skills.js'
 
-const HOOK_KEY = 'systematic_skill_tool_hooked'
-const SYSTEMATIC_MARKER = '__systematic_skill_tool__'
-
-interface HookState {
-  hookedTool: ToolDefinition | null
-  hookedDescription: string | null
-  initialized: boolean
-}
-
-const globalStore = globalThis as unknown as Record<string, HookState | undefined>
-
-function getHookState(): HookState {
-  let state = globalStore[HOOK_KEY]
-  if (state == null) {
-    state = {
-      hookedTool: null,
-      hookedDescription: null,
-      initialized: false,
-    }
-    globalStore[HOOK_KEY] = state
-  }
-  return state
-}
-
-export function isAlreadyHooked(): boolean {
-  const state = globalStore[HOOK_KEY]
-  return state != null && state.initialized
-}
-
-export function setHookedTool(hookedTool: ToolDefinition | null): void {
-  const state = getHookState()
-  state.hookedTool = hookedTool
-  state.initialized = true
-  if (hookedTool != null) {
-    state.hookedDescription =
-      typeof hookedTool.description === 'string' ? hookedTool.description : null
-  } else {
-    state.hookedDescription = null
-  }
-}
-
-export function getHookedTool(): ToolDefinition | null {
-  return getHookState().hookedTool
-}
-
-export function resetHookState(): void {
-  delete globalStore[HOOK_KEY]
-}
-
-function formatSkillsXml(skills: SkillInfo[]): string {
-  if (skills.length === 0) return ''
-
-  const skillsXml = skills
-    .map((skill) => {
-      const lines = [
-        '  <skill>',
-        `    <name>systematic:${skill.name}</name>`,
-        `    <description>${skill.description}</description>`,
-      ]
-      lines.push('  </skill>')
-      return lines.join('\n')
-    })
-    .join('\n')
-
-  return `<available_skills>\n${skillsXml}\n</available_skills>`
-}
-
-function mergeDescriptions(
-  baseDescription: string,
-  hookedDescription: string | null,
-  systematicSkillsXml: string
-): string {
-  if (hookedDescription == null || hookedDescription.trim() === '') {
-    return `${baseDescription}\n\n${systematicSkillsXml}`
-  }
-
-  const availableSkillsMatch = hookedDescription.match(
-    /<available_skills>([\s\S]*?)<\/available_skills>/
-  )
-
-  if (availableSkillsMatch) {
-    const existingSkillsContent = availableSkillsMatch[1]
-    const systematicSkillsContent = systematicSkillsXml
-      .replace('<available_skills>', '')
-      .replace('</available_skills>', '')
-      .trim()
-
-    const mergedContent = `<available_skills>\n${systematicSkillsContent}\n${existingSkillsContent}</available_skills>`
-    return hookedDescription.replace(
-      /<available_skills>[\s\S]*?<\/available_skills>/,
-      mergedContent
-    )
-  }
-
-  return `${hookedDescription}\n\n${systematicSkillsXml}`
-}
 
 function wrapSkillContent(skillPath: string, content: string): string {
   const skillDir = path.dirname(skillPath)
@@ -125,7 +30,7 @@ export function createSkillTool(options: SkillToolOptions): ToolDefinition {
   const { bundledSkillsDir, disabledSkills } = options
 
   const getSystematicSkills = (): SkillInfo[] => {
-    return findSkillsInDir(bundledSkillsDir, 'bundled', 3)
+    return findSkillsInDir(bundledSkillsDir)
       .filter((s) => !disabledSkills.includes(s.name))
       .sort((a, b) => a.name.localeCompare(b.name))
   }
@@ -139,17 +44,12 @@ export function createSkillTool(options: SkillToolOptions): ToolDefinition {
 Skills provide specialized knowledge and step-by-step guidance.
 Use this when a task matches an available skill's description.`
 
-    const hookState = getHookState()
-    return mergeDescriptions(
-      baseDescription,
-      hookState.hookedDescription,
-      systematicXml
-    )
+    return `${baseDescription}\n\n${systematicXml}`
   }
 
   let cachedDescription: string | null = null
 
-  const toolDef = tool({
+  return tool({
     get description() {
       if (cachedDescription == null) {
         cachedDescription = buildDescription()
@@ -192,27 +92,10 @@ ${wrapped}`
         }
       }
 
-      const hookedTool = getHookedTool()
-      if (hookedTool != null && typeof hookedTool.execute === 'function') {
-        try {
-          return await (hookedTool.execute as (args: { name: string }) => Promise<string>)(args)
-        } catch {
-          // Fallback failed, continue to error below
-        }
-      }
-
       const availableSystematic = skills.map((s) => `systematic:${s.name}`)
       throw new Error(
         `Skill "${requestedName}" not found. Available systematic skills: ${availableSystematic.join(', ')}`
       )
     },
   })
-
-  Object.defineProperty(toolDef, SYSTEMATIC_MARKER, {
-    value: true,
-    enumerable: false,
-    writable: false,
-  })
-
-  return toolDef
 }
