+++ tests/unit/skills.test.ts
@@ -2,9 +2,12 @@ import { afterEach, beforeEach, describe, expect, test } from 'bun:test'
 import fs from 'node:fs'
 import os from 'node:os'
 import path from 'node:path'
-import * as skillsCore from '../../src/lib/skills-core.ts'
+import * as agents from '../../src/lib/agents.ts'
+import * as commands from '../../src/lib/commands.ts'
+import * as frontmatter from '../../src/lib/frontmatter.ts'
+import * as skills from '../../src/lib/skills.ts'
 
-describe('skills-core', () => {
+describe('skills', () => {
   let testDir: string
 
   beforeEach(() => {
@@ -26,7 +29,7 @@ description: A test skill
 ---
 # Skill Content`,
       )
-      const result = skillsCore.extractFrontmatter(filePath)
+      const result = skills.extractFrontmatter(filePath)
       expect(result.name).toBe('test-skill')
       expect(result.description).toBe('A test skill')
     })
@@ -34,7 +37,7 @@ description: A test skill
     test('returns empty strings for missing frontmatter', () => {
       const filePath = path.join(testDir, 'test.md')
       fs.writeFileSync(filePath, '# Just a heading\nSome content')
-      const result = skillsCore.extractFrontmatter(filePath)
+      const result = skills.extractFrontmatter(filePath)
       expect(result.name).toBe('')
       expect(result.description).toBe('')
     })
@@ -47,26 +50,9 @@ description: A test skill
 ---
 # Content`,
       )
-      const result = skillsCore.extractFrontmatter(filePath)
+      const result = skills.extractFrontmatter(filePath)
       expect(result.name).toBe('')
       expect(result.description).toBe('')
-    })
-  })
-
-  describe('stripFrontmatter', () => {
-    test('removes frontmatter from content', () => {
-      const content = `---
-name: test
----
-# Content Here`
-      const result = skillsCore.stripFrontmatter(content)
-      expect(result).toBe('# Content Here')
-    })
-
-    test('returns content unchanged if no frontmatter', () => {
-      const content = '# No frontmatter'
-      const result = skillsCore.stripFrontmatter(content)
-      expect(result).toBe('# No frontmatter')
     })
   })
 
@@ -83,15 +69,14 @@ description: Test skill
 # My Skill`,
       )
 
-      const skills = skillsCore.findSkillsInDir(testDir, 'bundled')
-      expect(skills).toHaveLength(1)
-      expect(skills[0].name).toBe('my-skill')
-      expect(skills[0].sourceType).toBe('bundled')
+      const result = skills.findSkillsInDir(testDir)
+      expect(result).toHaveLength(1)
+      expect(result[0].name).toBe('my-skill')
     })
 
     test('returns empty array for non-existent directory', () => {
-      const skills = skillsCore.findSkillsInDir('/nonexistent/path', 'bundled')
-      expect(skills).toEqual([])
+      const result = skills.findSkillsInDir('/nonexistent/path')
+      expect(result).toEqual([])
     })
 
     test('uses directory name if no name in frontmatter', () => {
@@ -99,10 +84,41 @@ description: Test skill
       fs.mkdirSync(skillDir)
       fs.writeFileSync(path.join(skillDir, 'SKILL.md'), '# Just content')
 
-      const skills = skillsCore.findSkillsInDir(testDir, 'bundled')
-      expect(skills).toHaveLength(1)
-      expect(skills[0].name).toBe('unnamed-skill')
+      const result = skills.findSkillsInDir(testDir)
+      expect(result).toHaveLength(1)
+      expect(result[0].name).toBe('unnamed-skill')
     })
+  })
+})
+
+describe('frontmatter', () => {
+  describe('stripFrontmatter', () => {
+    test('removes frontmatter from content', () => {
+      const content = `---
+name: test
+---
+# Content Here`
+      const result = frontmatter.stripFrontmatter(content)
+      expect(result).toBe('# Content Here')
+    })
+
+    test('returns content unchanged if no frontmatter', () => {
+      const content = '# No frontmatter'
+      const result = frontmatter.stripFrontmatter(content)
+      expect(result).toBe('# No frontmatter')
+    })
+  })
+})
+
+describe('agents', () => {
+  let testDir: string
+
+  beforeEach(() => {
+    testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'systematic-test-'))
+  })
+
+  afterEach(() => {
+    fs.rmSync(testDir, { recursive: true, force: true })
   })
 
   describe('findAgentsInDir', () => {
@@ -116,11 +132,22 @@ description: Test agent
 # My Agent`,
       )
 
-      const agents = skillsCore.findAgentsInDir(testDir, 'bundled')
-      expect(agents).toHaveLength(1)
-      expect(agents[0].name).toBe('my-agent')
-      expect(agents[0].sourceType).toBe('bundled')
+      const result = agents.findAgentsInDir(testDir)
+      expect(result).toHaveLength(1)
+      expect(result[0].name).toBe('my-agent')
     })
+  })
+})
+
+describe('commands', () => {
+  let testDir: string
+
+  beforeEach(() => {
+    testDir = fs.mkdtempSync(path.join(os.tmpdir(), 'systematic-test-'))
+  })
+
+  afterEach(() => {
+    fs.rmSync(testDir, { recursive: true, force: true })
   })
 
   describe('findCommandsInDir', () => {
@@ -134,18 +161,17 @@ description: Test command
 # Test Command`,
       )
 
-      const commands = skillsCore.findCommandsInDir(testDir, 'bundled')
-      expect(commands).toHaveLength(1)
-      expect(commands[0].name).toBe('/sys-test')
-      expect(commands[0].sourceType).toBe('bundled')
+      const result = commands.findCommandsInDir(testDir)
+      expect(result).toHaveLength(1)
+      expect(result[0].name).toBe('/sys-test')
     })
 
     test('handles non-sys commands', () => {
       fs.writeFileSync(path.join(testDir, 'other-cmd.md'), '# Other')
 
-      const commands = skillsCore.findCommandsInDir(testDir, 'bundled')
-      expect(commands).toHaveLength(1)
-      expect(commands[0].name).toBe('/other-cmd')
+      const result = commands.findCommandsInDir(testDir)
+      expect(result).toHaveLength(1)
+      expect(result[0].name).toBe('/other-cmd')
     })
   })
 })
