+++ tsconfig.json
@@ -1,19 +1,19 @@
 {
   "compilerOptions": {
-    "target": "ES2022",
+    "target": "ESNext",
     "module": "ESNext",
     "moduleResolution": "bundler",
-    "lib": ["ES2022"],
+    "lib": ["ESNext"],
     "types": ["bun-types", "node"],
     "strict": true,
     "skipLibCheck": true,
-    "declaration": false,
-    "outDir": "./dist",
-    "rootDir": "./src",
+    "declaration": true,
+    "declarationDir": "dist",
+    "outDir": "dist",
+    "rootDir": "src",
     "esModuleInterop": true,
     "resolveJsonModule": true,
-    "isolatedModules": true,
-    "noEmit": true
+    "isolatedModules": true
   },
   "include": ["src/**/*"],
   "exclude": ["node_modules", "dist", "lib", ".opencode"]
